+++ csrc/opcua_types/sopc_encoder.h
@@ -22,7 +22,7 @@
 #include "sopc_builtintypes.h"
 #include "sopc_encodeabletype.h"
 
-#define SWAP_2_BYTES(x) (((x) &0x00FF) << 8 | ((x) &0xFF00) >> 8)
+#define SWAP_2_BYTES(x) (uint16_t)(((x) & (uint16_t) 0x00FF) << 8 | ((x) & (uint16_t) 0xFF00) >> 8)
 #define SWAP_3_BYTES(x) (((x) &0x0000FF) << 16 | ((x) &0x00FF00) | ((x) &0xFF0000) >> 16)
 #define SWAP_4_BYTES(x) \
     (((x) &0x000000FF) << 24 | ((x) &0x0000FF00) << 8 | ((x) &0xFF000000) >> 24 | ((x) &0x00FF0000) >> 8)
