+++ csrc/services/b2c/request_handle_bs.c
@@ -85,7 +85,7 @@ void request_handle_bs__client_fresh_req_handle(const constants__t_msg_type_i re
     {
         while (false == noHandleAvailable && *request_handle_bs__request_handle == constants__c_request_handle_indet)
         {
-            cpt = (cpt + 1) % (SOPC_MAX_PENDING_REQUESTS + 1);
+            cpt = (uint16_t)((cpt + 1) % (SOPC_MAX_PENDING_REQUESTS + 1));
             if (cpt == startedIdx)
             {
                 // Note: startedIdx content is never tested (simplest implem)
